refactor(mcp): modularize discovery service
