+++ src/domain/services/mcp-discovery/capability-detector.ts
@@ -1,0 +1,10 @@
+import { MCPServerProfile } from './discovery-types.js';
+
+/**
+ * Determines server capabilities and feature support.
+ */
+export class CapabilityDetector {
+  detectCapabilities(_server: MCPServerProfile): void {
+    // TODO: Implement capability detection logic
+  }
+}
