+++ src/domain/services/mcp-discovery/failover-coordinator.ts
@@ -1,0 +1,10 @@
+import { MCPServerProfile } from './discovery-types.js';
+
+/**
+ * Coordinates automatic failover handling.
+ */
+export class FailoverCoordinator {
+  handleFailure(_failedServer: MCPServerProfile): void {
+    // TODO: Implement failover coordination
+  }
+}
