+++ src/domain/services/mcp-discovery/health-monitor.ts
@@ -1,0 +1,10 @@
+import { MCPServerProfile } from './discovery-types.js';
+
+/**
+ * Performs periodic health checks on MCP servers.
+ */
+export class HealthMonitor {
+  async checkHealth(_server: MCPServerProfile): Promise<void> {
+    // TODO: Implement health check logic
+  }
+}
