+++ src/domain/services/mcp-discovery/index.ts
@@ -1,0 +1,10 @@
+export * from './discovery-types.js';
+export * from './server-scanner.js';
+export * from './capability-detector.js';
+export * from './registry-manager.js';
+export * from './health-monitor.js';
+export * from './connection-manager.js';
+export * from './load-balancer.js';
+export * from './failover-coordinator.js';
+export * from './metrics-collector.js';
+export * from './mcp-discovery-service.js';
