+++ src/domain/services/mcp-discovery/metrics-collector.ts
@@ -1,0 +1,10 @@
+import { MCPServerProfile } from './discovery-types.js';
+
+/**
+ * Collects metrics related to server discovery and performance.
+ */
+export class MetricsCollector {
+  record(_server: MCPServerProfile, _metric: string, _value: number): void {
+    // TODO: Implement metrics collection
+  }
+}
